requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Mail==0.9.1
WTForms==3.0.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
MarkupSafe==2.1.3
SQLAlchemy==2.0.21
Alembic==1.12.0
email-validator==2.0.0
reportlab==4.0.4
openpyxl==3.1.2
python-dateutil==2.8.2